chore: create a completion module
When we moved to `lspower`, we kinda just crammed everything into
`src/server.rs` so we knew what we needed and what we didn't. This patch
moves the completion code out of `server.rs` and into its own module.
Additionally, it shifts some imports around. No code changes were made;
this is all moved code.
